sender/requirements.txt
Jinja2==3.1.2
pydantic[dotenv]==1.10.4
psycopg2-binary==2.9.5
pika==1.3.1